Added different scales!
Added support for different scales. So basically we now have pentatonic
and blues, too.

TODO: every mode works with every scale now, but some combinations
don’t really make sense because of the way the grade of a mode is
calculated. To be honest though, I don’t really know if the concept of
“Aeolian Pentatonic” or whatever even applies…