Mode grades
So the grade of a mode (this is probably an inaccurate term) should be
based on the major scale, so I’m hard coding that in for now.

If nothing else, now things like “A Pentatonic Aeolian” work as
expected. Although I’m still not sure the concept even makes sense.

Still TODO: some scales/modes definitely don’t make sense. The
Pentatonic Locrian doesn’t even have the root in it, technically. How
to make these combinations illegal?