Added relative modes!
Also printing the list of relative modes! It really only works with the
major scale though, in the sense that the results for other scales
doesn’t really make sense.

Also there’s a pretty obvious pattern going on in the printing
functions that I’ll get around to extract soon.